+++ Source/Scale/MemberUpdater.cs
@@ -70,13 +70,13 @@ namespace TweakScale
                             }
                             return acc;
                         }, ref debuglog);
-                    Tools.LogWf(debuglog);
+                    Log.warn(debuglog);
                 }
             }
 
             if (property != null && property.GetIndexParameters().Length > 0)
             {
-                Tools.LogWf("Property {0} on {1} requires indices, which TweakScale currently does not support.", name, objectType.Name);
+                Log.warn("Property {0} on {1} requires indices, which TweakScale currently does not support.", name, objectType.Name);
                 return null;
             }
             if (field == null && property == null)
@@ -84,8 +84,9 @@ namespace TweakScale
                 // handle special cases
                 if ((obj is PartModule) && (name == "inputResources" || name == "outputResources"))
                     return Create((obj as PartModule).resHandler, name);
-
-                Tools.LogWf("No valid member found for {0} in {1}", name, objectType.Name);
+                    
+                string partname = (obj is PartModule) ? (obj as PartModule).part.name : "<unk>";
+                Log.warn("No valid member found for {0} in {1} for {2}", name, objectType.Name, partname);
 
                 return null;
             }