+++ CHANGE_LOG.md
@@ -1,5 +1,32 @@
 # TweakScale :: Change Log
 
+* 2019-0804: 2.4.3.2 (Lisias) for KSP >= 1.4.1
+	+ This is an Emergencial Release due an Emergencial Release due an Emergencial Release. I love recursion, don't you? :P
+	+ Closing or reworking the following issues:
+		- [#65](https://github.com/net-lisias-ksp/TweakScale/issues/65) Support for new Nertea's Cryo Engines
+			- Thanks to [friznet](https://github.com/friznit) and [marr75](https://github.com/marr75).
+		- Fixing some tyops :P on Logging and Dialog Boxes.
+* 2019-0725: 2.4.3.1 (Lisias) for KSP >= 1.4.1
+	+ This is an emergencial Release due a Emergencial Release. :P
+	+ Adding KSPe Light facilites:
+		- Logging
+	+ Closing or reworking the following issues:
+		- [#31](https://github.com/net-lisias-ksp/TweakScale/issues/31) Preventing being ran over by other mods
+			- A misbehaviour on detecting the misbehaviour :) was fixed. 
+		- [#47](https://github.com/net-lisias-ksp/TweakScale/issues/47) Count failed Sanity Checks as a potential problem. Warn user.
+		- [#48](https://github.com/net-lisias-ksp/TweakScale/issues/48) Backport the Heterodox Logging system into Orthodox (using KSPe.Light)
+		- [#49](https://github.com/net-lisias-ksp/TweakScale/issues/49) Check the Default patches for problems due wildcard!
+		- [#50](https://github.com/net-lisias-ksp/TweakScale/issues/50) Check the patches for currently supported Add'Ons
+			- ModuleGeneratorExtended Behaviour 
+		- [#51](https://github.com/net-lisias-ksp/TweakScale/issues/51) Implement a "Cancel" button when Actions are given to MessageBox
+			- Yeah. Doing it right this time. 
+		- [#54](https://github.com/net-lisias-ksp/TweakScale/issues/54) [ERR ***FATAL*** link provided in KSP.log links to 404
+			- "Typo maldito, typo maldito - tralálálálálálá"
+		- [#56](https://github.com/net-lisias-ksp/TweakScale/issues/56) "Breaking Parts" patches
+		- [#57](https://github.com/net-lisias-ksp/TweakScale/issues/57) Implement Warning Dialogs
+			- Warnings about Overrules, parts that couldn't be 			- Doing it right this time! 
+checked and parts with TweakScale support withdrawn.
+		- [#58](https://github.com/net-lisias-ksp/TweakScale/issues/58) Mk4 System Patch (addendum)
 * 2019-0608: 2.4.3.0 (Lisias) for KSP >= 1.4.1
 	+ This is an emergencial Release due a Show Stopper issue (see Issue #34 below) with some new features.
 	+ Adding features: